colibri_vf: change toradex config block offset type from hex to int
Allow specifying a negative number as the Toradex config block offset
by changing its type from hex to int. This is required in the eMMC case
where this is used to force a location at the end of respective
hardware boot partition.

While at it fix SPL compilation by not compiling configblock handling
in this case for now.