ARM: vf610: support L2 cache for Vybrid
Support L2 cache which is available on for the Cortex-A5 core on
some Vybrid SoCs (VFx1x series). Vybrid seems to use revision r3p2
of the PL310 cache controller. To enable the cache controller use
this two configuration symbols:
 #define CONFIG_SYS_L2_PL310
 #define CONFIG_SYS_PL310_BASE CA5_L2C_BASE_ADDR

Note: U-Boot is currently not able to detect the existence of the
L2 cache at runtime. Hence enabling these configurations leads to
a crash on SoCs which do not have a L2 cache (e.g. VF500 on
Colibri VF50)