MLK-11367 imx: mx6ul: Add cma bootargs for 6ul 9x9 board
Since the i.MX6ul 9x9 evk board only has 256MB LPDDR2, while the
CMA size used in kernel configuration is 320MB, so we have to set
another value for the 9x9 evk board.

This patch sets the cma=96M bootargs to in uboot. So the kernel will
overwrite to use the new value.

Signed-off-by: Ye.Li <B37916@freescale.com>