colibri_vf: read kernel and device tree from static UBI volumes
With the next release our update scripts write the kernel and
device tree in seperate UBI volumes. This allows to use a lot
less UBI/UBIFS support in U-Boot, which should lower the risk
of hitting bugs in this area. Boot times are also slightly
faster (measurements showed an improvement of ~150ms).

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>