apalis_tk1: fix l4t booting by removing noatime mount option
Get rid of the noatime mount option as L4T resp. Ubuntu does not seem to like
that one at all and rather fails booting as follows:

 * Starting NVIDIA specific init script                                  [fail]
...
 * Starting NSM status monitor                                           [ OK ]
The disk drive for / is not ready yet or not present.
keys:Continue to wait, or Press S to skip mounting or M for manual recovery
 * Starting NSM status monitor                                           [ OK ]
...
 * Starting NSM status monitor                                           [fail]
 * Stopping NSM status monitor                                           [ OK ]
 * Starting Bridge socket events into upstart                            [ OK ]

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>